Add files via upload
This file is the claims severity dataset - it pulls in the corresponding policy level data from the earned dataset by the (what I believe is a unique tag) combination of TENURE and EXPOSURE_ID

because this represents the TIMING (tenure) + PET LABEL (exposure_id represents different insured "risks")

this severity dataset then needs to be treated before it is merged back into the policy level dataset. in my opinion, the policy level dataset can become 1 total dataset that contains both claim severity and frequency information. the final dataset may be a time series, monthly, of policies, and then there would be a count of # claims in that month, so maybe a bunch of 0s, and then some +ve number of claims for other months etc. when the policy claimed. hence the severity dataset must be treated before merged back into the policy level dataset. 

there are a few things to consider then within the tagged (tagged meaning the unique ID was used) severity dataset. 

1. how do we deal with claims from same pet (exposure ID), same condition? 
1. a) how do we deal with same pet, same condition, same tenure (same month)
1. b) how do we deal with same pet, same condition, different tenures (different months)

point 1 is all about how to concatenate/deal with claims related to the same risk, because this has importance when linking it back to a frequency dataset. 

for an insured pet of the same condition, eg 1 exposure ID, one condition like "illness", but multiple claims across time. One thought I had was 
- concatenate all claims into 1 row. New variables: flag for multiple claims, date of first claim, last claim, difference in smallest and largest tenure, # of claims made overall 

you would then repeat for the other conditions if the same pet claimed for different conditions like "gastrointestinal" etc. 

or you could run this procedure only if the claims occur in the same month, so you have different rows for the 1 exposure ID and condition eg "illness" but different months, and then create a new variable that flags this pet had more claims in other months. 


2. we need to back out the total claim amount calculation. this calculation depends on how you proceed with point 1. because you would need to apply the calculation claims_paid / contribution rate + excess / 1.1 ---- but for same exposure ID, same condition eg "illness", only the first claim has an excess applied. we need to identify the EARLIEST claim for EXPOSURE_ID and CONDITION. only that one gets the excess applied. 





once the severity dataset is treated, it can be merged back into the policy level dataset using the same unique ID.